Add class Component, update Task and TaskEdit classes
